perf(descriptions): split type.ts file (#2447)
* perf(descriptions): split type.ts file

* fix: add type
