+++ packages/descriptions/src/descriptions-row.vue
@@ -30,7 +30,8 @@
 <script lang="ts">
 import { defineComponent, inject } from 'vue'
 import DescriptionsCell from './descriptions-cell'
-import { elDescriptionsKey } from './descriptions.type'
+import { elDescriptionsKey } from './token'
+
 import type { IDescriptionsInject } from './descriptions.type'
 
 export default defineComponent({
