+++ packages/descriptions/src/index.vue
@@ -26,7 +26,7 @@ import { computed, defineComponent, provide } from 'vue'
 import { isValidComponentSize } from '@element-plus/utils/validators'
 import DescriptionsRow from './descriptions-row.vue'
 import { useGlobalConfig } from '@element-plus/utils/util'
-import { elDescriptionsKey } from './descriptions.type'
+import { elDescriptionsKey } from './token'
 
 import type { PropType } from 'vue'
