+++ packages/descriptions/src/token.ts
@@ -1,0 +1,4 @@
+import { InjectionKey } from 'vue'
+import { IDescriptionsInject } from './descriptions.type'
+
+export const elDescriptionsKey: InjectionKey<IDescriptionsInject> = 'elDescriptions' as any
